pyproject.toml
[project]
name = "ozbe14-miniproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.119.1",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "aerich>=0.9.2",
    "pytest>=8.4.2",
    "ruff>=0.14.1",
]
prod = [
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]

[tool.aerich]
tortoise_orm = "app.db.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
